+++ drive/platform-implementations/esp8266_mission_control.cpp
@@ -1,5 +1,3 @@
-
-#include "../src/mission_control.hpp"
 #include <array>
 #include <cinttypes>
 #include <string_view>
@@ -11,38 +9,25 @@
 #include <libhal-util/timeout.hpp>
 #include <libhal/timeout.hpp>
 
+#include "../include/mission_control.hpp"
+#include "../include/esp8266_mission_control.hpp"
+
 namespace sjsu::drive {
 
-class esp8266_mission_control : public mission_control
-{
-public:
-  static hal::result<esp8266_mission_control> create(
-    hal::esp8266::at& p_esp8266,
-    hal::serial& p_console,
-    const std::string_view p_ssid,
-    const std::string_view p_password,
-    const hal::esp8266::at::socket_config& p_config,
-    const std::string_view p_ip,
-    hal::timeout auto& p_timeout,
-    std::span<hal::byte> p_buffer,
-    std::string_view p_get_request)
+  esp8266_mission_control::http_header_parser_t esp8266_mission_control::new_http_header_parser()
   {
-    esp8266_mission_control esp_mission_control =
-      esp8266_mission_control(p_esp8266,
-                              p_console,
-                              p_ssid,
-                              p_password,
-                              p_config,
-                              p_ip,
-                              p_buffer,
-                              p_get_request);
-    HAL_CHECK(esp_mission_control.establish_connection(p_timeout));
+    using namespace std::literals;
 
-    return esp_mission_control;
+    return esp8266_mission_control::http_header_parser_t{
+      .find_header_start = hal::stream_find(hal::as_bytes("HTTP/1.1 "sv)),
+      .find_content_length =
+        hal::stream_find(hal::as_bytes("Content-Length: "sv)),
+      .parse_content_length = hal::stream_parse<std::uint32_t>(),
+      .find_end_of_header = hal::stream_find(hal::as_bytes("\r\n\r\n"sv)),
+    };
   }
 
-private:
-  esp8266_mission_control(hal::esp8266::at& p_esp8266,
+  esp8266_mission_control::esp8266_mission_control(hal::esp8266::at& p_esp8266,
                           hal::serial& p_console,
                           const std::string_view p_ssid,
                           const std::string_view p_password,
@@ -60,12 +45,116 @@ private:
     , m_get_request(p_get_request)
     , m_fill_payload(hal::stream_fill(m_buffer))
     , m_http_header_parser(new_http_header_parser())
+    , m_buffer_len(0)
   {
-    m_buffer_len = 0;
   }
 
-  hal::result<mc_commands> impl_get_command(
-    hal::function_ref<hal::timeout_function> p_timeout) override
+  hal::status esp8266_mission_control::establish_connection(hal::timeout auto& p_timeout)
+  {
+    connection_state state = connection_state::check_ap_connection;
+
+    while (state != connection_state::connection_established) {
+      switch (state) {
+        case connection_state::check_ap_connection:
+          hal::print(*m_console, "Checking if AP \"");
+          hal::print(*m_console, m_ssid);
+          hal::print(*m_console, "\" is connected... ");
+          if (HAL_CHECK(m_esp8266->is_connected_to_ap(p_timeout))) {
+            state = connection_state::check_server_connection;
+            hal::print(*m_console, "Connected!\n");
+          } else {
+            state = connection_state::connecting_to_ap;
+            hal::print(*m_console, "NOT Connected!\n");
+          }
+          break;
+        case connection_state::connecting_to_ap:
+          hal::print(*m_console, "Connecting to AP: \"");
+          hal::print(*m_console, m_ssid);
+          hal::print(*m_console, "\" ...\n");
+          HAL_CHECK(m_esp8266->connect_to_ap(m_ssid, m_password, p_timeout));
+          state = connection_state::set_ip_address;
+          break;
+        case connection_state::set_ip_address:
+          if (!m_ip.empty()) {
+            hal::print(*m_console, "Setting IP Address to: ");
+            hal::print(*m_console, m_ip);
+            hal::print(*m_console, " ...\n");
+            HAL_CHECK(m_esp8266->set_ip_address(m_ip, p_timeout));
+          }
+          state = connection_state::check_server_connection;
+          break;
+        case connection_state::check_server_connection:
+          hal::print(*m_console, "Checking if server \"");
+          hal::print(*m_console, m_config.domain);
+          hal::print(*m_console, "\" is connected... \n");
+          if (HAL_CHECK(m_esp8266->is_connected_to_server(p_timeout))) {
+            state = connection_state::connection_established;
+            hal::print(*m_console, "Connected!\n");
+          } else {
+            state = connection_state::connecting_to_server;
+            hal::print(*m_console, "NOT Connected!\n");
+          }
+          break;
+        case connection_state::connecting_to_server:
+          hal::print(*m_console, "Connecting to server: \"");
+          hal::print(*m_console, m_config.domain);
+          hal::print(*m_console, "\" ...\n");
+          HAL_CHECK(m_esp8266->connect_to_server(m_config, p_timeout));
+          hal::print(*m_console, "connected\n");
+          state = connection_state::check_server_connection;
+          break;
+        case connection_state::connection_established:
+          // Do nothing, allow next iteration to break while loop
+          hal::print(*m_console, "Succesfully Connected.");
+          break;
+        default:
+          state = connection_state::connecting_to_ap;
+      }
+    }
+
+    return hal::success();
+  }
+
+  std::string_view esp8266_mission_control::to_string_view(std::span<const hal::byte> p_span)
+  {
+    return std::string_view(reinterpret_cast<const char*>(p_span.data()),
+                            p_span.size());
+  }
+
+    void esp8266_mission_control::parse_commands() {
+
+    auto result = to_string_view(m_command_buffer);
+    static constexpr int expected_number_of_arguments = 6;
+    mc_commands commands;
+
+    int actual_arguments = sscanf(result.data(),
+                                  kResponseBodyFormat,
+                                  &commands.heartbeat_count,
+                                  &commands.is_operational,
+                                  &commands.wheel_orientation,
+                                  &commands.mode,
+                                  &commands.speed,
+                                  &commands.angle);
+    if (actual_arguments != expected_number_of_arguments) {
+      hal::print<200>(*m_console,
+                      "Received %d arguments, expected %d\n",
+                      actual_arguments,
+                      expected_number_of_arguments);
+    }
+    // hal::print<200>(*m_console,
+    //                   "HB: %d\t, IO %d\t, WO: %d\t, DM: %c\t, Speed: %d\n, Angle: %d\n",
+    //                   commands.heartbeat_count,
+    //                   commands.is_operational,
+    //                   commands.wheel_orientation,
+    //                   commands.mode,
+    //                   commands.speed,
+    //                   commands.angle
+    //                   );
+    m_commands = commands;
+  }
+
+  hal::result<mission_control::mc_commands> esp8266_mission_control::impl_get_command(
+    hal::function_ref<hal::timeout_function> p_timeout)
   {
     using namespace std::literals;
 
@@ -73,7 +162,6 @@ private:
     if (m_write_error) {
       hal::print(*m_console, "Reconnecting...\n");
       // Wait 1s before attempting to reconnect
-
       auto result = establish_connection(p_timeout);
       if (!result) {
         hal::print(*m_console, "Failure!!!\n");
@@ -158,159 +246,29 @@ private:
     return m_commands;
   }
 
-  void parse_commands() {
+    hal::result<esp8266_mission_control> esp8266_mission_control::create(
+        hal::esp8266::at& p_esp8266,
+        hal::serial& p_console,
+        const std::string_view p_ssid,
+        const std::string_view p_password,
+        const hal::esp8266::at::socket_config& p_config,
+        const std::string_view p_ip,
+        hal::timeout auto& p_timeout,
+        std::span<hal::byte> p_buffer,
+        std::string_view p_get_request)
+  {
+    esp8266_mission_control esp_mission_control =
+      esp8266_mission_control(p_esp8266,
+                              p_console,
+                              p_ssid,
+                              p_password,
+                              p_config,
+                              p_ip,
+                              p_buffer,
+                              p_get_request);
+    HAL_CHECK(esp_mission_control.establish_connection(p_timeout));
 
-    auto result = to_string_view(m_command_buffer);
-    static constexpr int expected_number_of_arguments = 6;
-    mc_commands commands;
-
-    int actual_arguments = sscanf(result.data(),
-                                  kResponseBodyFormat,
-                                  &commands.heartbeat_count,
-                                  &commands.is_operational,
-                                  &commands.wheel_orientation,
-                                  &commands.mode,
-                                  &commands.speed,
-                                  &commands.angle);
-    if (actual_arguments != expected_number_of_arguments) {
-      hal::print<200>(*m_console,
-                      "Received %d arguments, expected %d\n",
-                      actual_arguments,
-                      expected_number_of_arguments);
-    }
-    // hal::print<200>(*m_console,
-    //                   "HB: %d\t, IO %d\t, WO: %d\t, DM: %c\t, Speed: %d\n, Angle: %d\n",
-    //                   commands.heartbeat_count,
-    //                   commands.is_operational,
-    //                   commands.wheel_orientation,
-    //                   commands.mode,
-    //                   commands.speed,
-    //                   commands.angle
-    //                   );
-    m_commands = commands;
+    return esp_mission_control;
   }
 
-
-  std::string_view to_string_view(std::span<const hal::byte> p_span)
-  {
-    return std::string_view(reinterpret_cast<const char*>(p_span.data()),
-                            p_span.size());
-  }
-
-  enum class connection_state
-  {
-    check_ap_connection,
-    connecting_to_ap,
-    set_ip_address,
-    check_server_connection,
-    connecting_to_server,
-    connection_established,
-  };
-
-  [[nodiscard]] hal::status establish_connection(hal::timeout auto& p_timeout)
-  {
-    connection_state state = connection_state::check_ap_connection;
-
-    while (state != connection_state::connection_established) {
-      switch (state) {
-        case connection_state::check_ap_connection:
-          hal::print(*m_console, "Checking if AP \"");
-          hal::print(*m_console, m_ssid);
-          hal::print(*m_console, "\" is connected... ");
-          if (HAL_CHECK(m_esp8266->is_connected_to_ap(p_timeout))) {
-            state = connection_state::check_server_connection;
-            hal::print(*m_console, "Connected!\n");
-          } else {
-            state = connection_state::connecting_to_ap;
-            hal::print(*m_console, "NOT Connected!\n");
-          }
-          break;
-        case connection_state::connecting_to_ap:
-          hal::print(*m_console, "Connecting to AP: \"");
-          hal::print(*m_console, m_ssid);
-          hal::print(*m_console, "\" ...\n");
-          HAL_CHECK(m_esp8266->connect_to_ap(m_ssid, m_password, p_timeout));
-          state = connection_state::set_ip_address;
-          break;
-        case connection_state::set_ip_address:
-          if (!m_ip.empty()) {
-            hal::print(*m_console, "Setting IP Address to: ");
-            hal::print(*m_console, m_ip);
-            hal::print(*m_console, " ...\n");
-            HAL_CHECK(m_esp8266->set_ip_address(m_ip, p_timeout));
-          }
-          state = connection_state::check_server_connection;
-          break;
-        case connection_state::check_server_connection:
-          hal::print(*m_console, "Checking if server \"");
-          hal::print(*m_console, m_config.domain);
-          hal::print(*m_console, "\" is connected... \n");
-          if (HAL_CHECK(m_esp8266->is_connected_to_server(p_timeout))) {
-            state = connection_state::connection_established;
-            hal::print(*m_console, "Connected!\n");
-          } else {
-            state = connection_state::connecting_to_server;
-            hal::print(*m_console, "NOT Connected!\n");
-          }
-          break;
-        case connection_state::connecting_to_server:
-          hal::print(*m_console, "Connecting to server: \"");
-          hal::print(*m_console, m_config.domain);
-          hal::print(*m_console, "\" ...\n");
-          HAL_CHECK(m_esp8266->connect_to_server(m_config, p_timeout));
-          hal::print(*m_console, "connected\n");
-          state = connection_state::check_server_connection;
-          break;
-        case connection_state::connection_established:
-          // Do nothing, allow next iteration to break while loop
-          hal::print(*m_console, "Succesfully Connected.");
-          break;
-        default:
-          state = connection_state::connecting_to_ap;
-      }
-    }
-
-    return hal::success();
-  }
-
-  struct http_header_parser_t
-  {
-    hal::stream_find find_header_start;
-    hal::stream_find find_content_length;
-    hal::stream_parse<std::uint32_t> parse_content_length;
-    hal::stream_find find_end_of_header;
-  };
-
-  http_header_parser_t new_http_header_parser()
-  {
-    using namespace std::literals;
-
-    return http_header_parser_t{
-      .find_header_start = hal::stream_find(hal::as_bytes("HTTP/1.1 "sv)),
-      .find_content_length =
-        hal::stream_find(hal::as_bytes("Content-Length: "sv)),
-      .parse_content_length = hal::stream_parse<std::uint32_t>(),
-      .find_end_of_header = hal::stream_find(hal::as_bytes("\r\n\r\n"sv)),
-    };
-  }
-
-  mc_commands m_commands{};
-  hal::esp8266::at* m_esp8266;
-  hal::serial* m_console;
-  std::string_view m_ssid;
-  std::string_view m_password;
-  const hal::esp8266::at::socket_config& m_config;
-  std::string_view m_ip;
-  std::span<hal::byte> m_buffer;
-  std::array<hal::byte, 128> m_command_buffer;
-  std::string_view m_get_request;
-  http_header_parser_t m_http_header_parser;
-  size_t m_buffer_len;
-  bool m_write_error = false;
-  bool m_header_finished = false;
-  bool m_read_complete = true;
-  hal::stream_fill m_fill_payload;
-  size_t m_content_length;
-  std::uint32_t m_missed_read = 0;
-};
 }  // namespace sjsu::arm
